Extend selectOption() to select text that partially match
The WebDriver module is intended to be a replacement of Selenium2
module. In the Selenium2 module selectOption($field, $option) would
select an option where the text of the option contained $option.
php-facebook's WebDriverSelect->selectByVisibleText() only selects
on full match. This patch updates the function to match the behavior
of Selenium2.