Ignore errors from lastInsertId()
lastInsertId() fails if the DB structure does not match Db module's assumption,
for example, a table without tablename_id_seq sequence in PostgreSQL.

Db module is to be simple, so it is enough to ignore the exception for
uncommon conditions.

When lastInsertId() is not available, cleaning up the record is also impossible.
So the row is not stored in $this->insertedIds.