Symfony2: Cache router
Router can be cached on Symfony2 framework. This option is disabled by default.
It can be enabled with:
modules:
  - Symfony2:
    cache_router: true
    ...

It's also provided a method to force cache invalidation on desire named
'invalidateCachedRouter()', to be used on the rare cases you want router to be
cached but need to manually invalidate cache before of after certain tests.