Minor improvements (#4225)
* Minor improvements

* Minor improvements

* Minor improvements

* Minor improvements

* Please double-check!

What was meant by WebDriver not being able to "perform HTTP requests"?

* Minor improvements

* Explained stopping after failed assertions

* Minor improvements

* Minor improvements

* Please double-check

Is it a separate transaction for each test? Or one transaction for all the tests?

Plus: Please explain how this is to be done: "If it’s not your required behavior, change the settings of the Db module for the current suite."

* Minor improvements

* Minor improvements

* Shortened the _pageProvider codeblock

Please explain: What's the advantage of having a *public* method for the data provider? Why would somebody want to have this?

* Minor improvements

* Some improvements