Ignore files when exporting package (#5527)
The [install file](https://github.com/Codeception/Codeception/archive/3.0.zip) downloaded by Composer is currently 1.1 MB.

These changes reduce it to ~0.4 MB, speeding up installs. Won't matter most of the time, CI servers etc. but noticeable when a developer is installing via a mobile tethered connection for example.

Codeception [is installed ~373,000 times per month](https://packagist.org/packages/codeception/codeception/stats), so there's also (as noted in one of the background links below) an assumed carbon reduction bonus in not transferring the ~261 GB difference every month, but I've no idea if that's even possible to measure! :-)

We use (and love!) Codeception daily, but someone more familiar with the package itself will hopefully be able to verify I've not excluded anything expected to be present on the developers machine after installing - and maybe even exclude some I haven't. I don't believe this to be a breaking change, since the files excluded are for developing Codeception, rather than to use it.

Background:

- https://madewithlove.be/gitattributes/
- https://www.reddit.com/r/PHP/comments/2jzp6k/i_dont_need_your_tests_in_my_production/