Do not truncate arguments for --html options
By running the `run --html`, we can have a report.html file that is generated.
However, the content of this file is truncated, thus making the report sometimes unusable.

This is a quick proposal but IMO it should not be merged as it is.
We should rather create an optional conf input (or whatever) to let the user chose if yes or not, he wants the truncate to happen, and how maximal characters he desires.

Cover the "too" long argument case.